Fix int in resolution test

+++ tests/test_objective_funs.py
@@ -1874,7 +1874,7 @@ class TestComputeScalarResolution:
         f = np.zeros_like(self.res_array, dtype=float)
         for i, res in enumerate(self.res_array):
             obj = ObjectiveFunction(
-                objective(coilset, grid=LinearGrid(N=5 + 3 * res)), use_jit=False
+                objective(coilset, grid=LinearGrid(N=int(5 + 3 * res))), use_jit=False
             )
             obj.build(verbose=0)
             f[i] = obj.compute_scalar(obj.x())
